refactor: transaction overview section
